requirement.txt
Django
djangorestframework
djangorestframework-simplejwt
python-decouple